requirements.txt
aiocqhttp==1.4.1
aiofiles==0.7.0
aiohttp==3.7.4.post0
altgraph==0.17
anyio==3.3.0
APScheduler==3.7.0
arrow==1.1.1
asgiref==3.4.1
async-lru==1.0.2
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
crypto==1.4.1
cryptography==3.4.7
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
easygui==0.98.2
et-xmlfile==1.0.1
fastapi==0.68.0
ffmpeg==1.4
future==0.18.2
goto-statement==1.2
h11==0.12.0
h2==3.2.0
h5py==3.3.0
hpack==3.0.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
Hypercorn==0.11.2
hyperframe==5.2.0
idna==2.7
imageio==2.9.0
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==1.0.1
kiwisolver==1.3.1
loguru==0.5.3
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.1
multidict==5.1.0
Naked==0.1.31
nb-cli==0.5.0
networkx==2.5.1
nonebot-adapter-cqhttp==2.0.0a14
nonebot-adapter-ding==2.0.0a14
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0a14
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
pefile==2019.4.18
Pillow==8.3.1
poyo==0.5.0
priority==2.0.0
prompt-toolkit==3.0.19
psutil==5.8.0
pyaes==1.6.1
pybase64==1.1.4
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pyfiglet==0.8.post1
pygame==2.0.1
pygtrie==2.4.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypinyin==0.42.0
Pyrogram==1.2.9
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.0
python-slugify==5.0.2
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
Quart==0.14.1
requests==2.20.0
rfc3986==1.5.0
scikit-learn==0.24.2
selenium==3.141.0
shellescape==3.8.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
starlette==0.14.2
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tomlkit==0.7.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.24.3
uvicorn==0.14.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.1
win32-setctime==1.0.3
wordcloud @ file:///D:/Program%20Files/Python39/Lib/site-packages/wordcloud-1.8.1-cp39-cp39-win_amd64.whl
wsproto==1.0.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zope.interface==5.4.0
